moved parts from pages to components
