+++ components/Liquidations.tsx
@@ -7,9 +7,9 @@ import React from "react";
 import { useEffect, useState } from "react";
 import { LiquidationQueueClient, LiquidationQueueQueryClient } from "../codegen/liquidation_queue/LiquidationQueue.client";
 import { ClaimsResponse, QueueResponse, SlotResponse } from "../codegen/liquidation_queue/LiquidationQueue.types";
-import { denoms, Prices } from ".";
+import { denoms, Prices } from "../pages";
 import { coins } from "@cosmjs/stargate";
-import Popup from "../components/Popup";
+import Popup from "./Popup";
 import { StabilityPoolClient, StabilityPoolQueryClient } from "../codegen/stability_pool/StabilityPool.client";
 import { PositionsQueryClient } from "../codegen/positions/Positions.client";
 import { NativeToken } from "../codegen/positions/Positions.types";
