+++ components/Vaults.tsx
@@ -8,8 +8,8 @@ import { Coin, coin, coins, parseCoins } from "@cosmjs/amino";
 import { StargateClient } from "@cosmjs/stargate";
 import { PositionsClient, PositionsQueryClient } from "../codegen/positions/Positions.client";
 import { Asset, Basket, BasketPositionsResponse, CollateralInterestResponse, InterestResponse, NativeToken, PositionResponse, RedeemabilityResponse } from "../codegen/positions/Positions.types";
-import { denoms, Prices } from ".";
-import Popup from "../components/Popup";
+import { denoms, Prices } from "../pages";
+import Popup from "./Popup";
 import Image from "next/image";
 import { relative } from "path";
 import { ReactJSXElement } from "@emotion/react/types/jsx-namespace";
